Update AdminMainOptions.vue
Hello Jordan,
I have been using your amazing plugin for a project on our website to refresh and reload certain aspects of a streaming page, I thought to add a few less seconds (5 and 10) to the time interval for the refresh. I would like for you to take a look and probably ok it so that I can use it for the rest of the project, thank you in anticipation.